refactor(web): 抽出浏览器相关帮助函数
